[Asset][FrameworkBundle] Add AssetBundle
The asset services move to the Asset component, which now provides its own bundle and the
"asset" configuration. The "framework.assets" configuration becomes an alias of it.

No inversion was needed: the only two readers outside the section, TwigBundle's ExtensionPass
and AssetMapper's RemoveMissingDependenciesPass, already ask the merged container for
"assets.packages" and "assets._default_package" rather than reading the config flag.

The autoconfiguration of PackageInterface, registered unconditionally by FrameworkExtension
until now, moves into the bundle and so follows the section's enabled flag.
