+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -145,7 +145,7 @@ class Configuration implements ConfigurationInterface
         $this->addRouterSection($rootNode);
         $this->addSessionSection($rootNode);
         $this->addRequestSection($rootNode);
-        $this->addAssetsSection($rootNode, $enableIfStandalone);
+        $this->addAssetsSection($rootNode);
         $this->addAssetMapperSection($rootNode);
         $this->addTranslatorSection($rootNode, $enableIfStandalone);
         $this->addValidationSection($rootNode, $enableIfStandalone);
@@ -550,82 +550,14 @@ class Configuration implements ConfigurationInterface
     /**
      * @param-immediately-invoked-callable $enableIfStandalone
      */
-    private function addAssetsSection(ArrayNodeDefinition $rootNode, callable $enableIfStandalone): void
+    private function addAssetsSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
             ->children()
-                ->arrayNode('assets')
-                    ->info('Assets configuration')
-                    ->{$enableIfStandalone('symfony/asset', Package::class)}()
-                    ->children()
-                        ->booleanNode('strict_mode')
-                            ->info('Throw an exception if an entry is missing from the manifest.json.')
-                            ->defaultFalse()
-                        ->end()
-                        ->scalarNode('version_strategy')->defaultNull()->end()
-                        ->scalarNode('version')->defaultNull()->end()
-                        ->scalarNode('version_format')->defaultValue('%%s?%%s')->end()
-                        ->scalarNode('json_manifest_path')->defaultNull()->end()
-                        ->scalarNode('base_path')->defaultValue('')->end()
-                        ->arrayNode('base_urls', 'base_url')
-                            ->requiresAtLeastOneElement()
-                            ->acceptAndWrap(['string'])
-                            ->prototype('scalar')->end()
-                        ->end()
-                    ->end()
-                    ->validate()
-                        ->ifTrue(static fn ($v) => isset($v['version_strategy']) && isset($v['version']))
-                        ->thenInvalid('You cannot use both "version_strategy" and "version" at the same time under "assets".')
-                    ->end()
-                    ->validate()
-                        ->ifTrue(static fn ($v) => isset($v['version_strategy']) && isset($v['json_manifest_path']))
-                        ->thenInvalid('You cannot use both "version_strategy" and "json_manifest_path" at the same time under "assets".')
-                    ->end()
-                    ->validate()
-                        ->ifTrue(static fn ($v) => isset($v['version']) && isset($v['json_manifest_path']))
-                        ->thenInvalid('You cannot use both "version" and "json_manifest_path" at the same time under "assets".')
-                    ->end()
-                    ->children()
-                        ->arrayNode('packages', 'package')
-                            ->normalizeKeys(false)
-                            ->useAttributeAsKey('name')
-                            ->prototype('array')
-                                ->children()
-                                    ->booleanNode('strict_mode')
-                                        ->info('Throw an exception if an entry is missing from the manifest.json.')
-                                        ->defaultFalse()
-                                    ->end()
-                                    ->scalarNode('version_strategy')->defaultNull()->end()
-                                    ->scalarNode('version')
-                                        ->beforeNormalization()
-                                            ->ifString()
-                                            ->then(static fn ($v) => '' === $v ? null : $v)
-                                        ->end()
-                                    ->end()
-                                    ->scalarNode('version_format')->defaultNull()->end()
-                                    ->scalarNode('json_manifest_path')->defaultNull()->end()
-                                    ->scalarNode('base_path')->defaultValue('')->end()
-                                    ->arrayNode('base_urls', 'base_url')
-                                        ->requiresAtLeastOneElement()
-                                        ->acceptAndWrap(['string'])
-                                        ->prototype('scalar')->end()
-                                    ->end()
-                                ->end()
-                                ->validate()
-                                    ->ifTrue(static fn ($v) => isset($v['version_strategy']) && isset($v['version']))
-                                    ->thenInvalid('You cannot use both "version_strategy" and "version" at the same time under "assets" packages.')
-                                ->end()
-                                ->validate()
-                                    ->ifTrue(static fn ($v) => isset($v['version_strategy']) && isset($v['json_manifest_path']))
-                                    ->thenInvalid('You cannot use both "version_strategy" and "json_manifest_path" at the same time under "assets" packages.')
-                                ->end()
-                                ->validate()
-                                    ->ifTrue(static fn ($v) => isset($v['version']) && isset($v['json_manifest_path']))
-                                    ->thenInvalid('You cannot use both "version" and "json_manifest_path" at the same time under "assets" packages.')
-                                ->end()
-                            ->end()
-                        ->end()
-                    ->end()
+                ->variableNode('assets')
+                    ->aliasOf('asset')
+                    ->treatFalseLike(['enabled' => false])
+                    ->treatTrueLike(['enabled' => true])
                 ->end()
             ->end()
         ;
