+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/asset_mapper_without_assets.php
@@ -2,5 +2,4 @@
 
 $container->loadFromExtension('framework', [
     'asset_mapper' => null,
-    'assets' => false,
 ]);
