--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/assets_version_strategy_as_service.php
@@ -1,8 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'assets' => [
-        'version_strategy' => 'assets.custom_version_strategy',
-        'base_urls' => 'http://cdn.example.com',
-    ],
-]);
