+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/legacy_assets.php
@@ -1,7 +1,5 @@
 <?php
 
 $container->loadFromExtension('framework', [
-    'assets' => [
-        'enabled' => false,
-    ],
+    'assets' => ['version' => 'v1'],
 ]);
