+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/asset_mapper_without_assets.yml
@@ -1,3 +1,2 @@
 framework:
     asset_mapper: ~
-    assets: false
