--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/assets.yml
@@ -1,34 +1,0 @@
-framework:
-    assets:
-        version: SomeVersionScheme
-        version_format: '%%s?version=%%s'
-        base_urls: http://cdn.example.com
-        packages:
-            images_path:
-                base_path: '/foo'
-            images:
-                version: 1.0.0
-                base_urls: ["http://images1.example.com", "http://images2.example.com"]
-            foo:
-                version: 1.0.0
-                version_format: '%%s-%%s'
-            bar:
-                base_urls: ["https://bar2.example.com"]
-            bar_version_strategy:
-                base_urls: ["https://bar_version_strategy.example.com"]
-                version_strategy: assets.custom_version_strategy
-            json_manifest_strategy:
-                json_manifest_path: '/path/to/manifest.json'
-            remote_manifest:
-                json_manifest_path: 'https://cdn.example.com/manifest.json'
-            var_manifest:
-                json_manifest_path: '%var_json_manifest_path%'
-            env_manifest:
-                json_manifest_path: '%env(env_manifest)%'
-            strict_manifest_strategy:
-                json_manifest_path: '/path/to/manifest.json'
-                strict_mode: true
-
-parameters:
-    var_json_manifest_path: 'https://cdn.example.com/manifest.json'
-    env(env_manifest): https://cdn.example.com/manifest.json
