--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/assets_version_strategy_as_service.yml
@@ -1,4 +1,0 @@
-framework:
-    assets:
-        version_strategy: assets.custom_version_strategy
-        base_urls: http://cdn.example.com
