+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/full.yml
@@ -33,8 +33,6 @@ framework:
         gc_divisor:      108
         gc_maxlifetime:  90000
         save_path:       /path/to/sessions
-    assets:
-        version: v1
     translator:
         enabled:  true
         fallback: fr
