+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/legacy_assets.yml
@@ -1,3 +1,3 @@
 framework:
     assets:
-        enabled: false
+        version: v1
