+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -42,7 +42,7 @@
         "phpdocumentor/reflection-docblock": "^5.2|^6.0",
         "phpstan/phpdoc-parser": "^1.0|^2.0",
         "seld/jsonlint": "^1.10",
-        "symfony/asset": "^7.4|^8.0",
+        "symfony/asset": "^8.2",
         "symfony/asset-mapper": "^8.2",
         "symfony/browser-kit": "^7.4|^8.0",
         "symfony/clock": "^7.4|^8.0",
@@ -88,6 +88,7 @@
         "doctrine/persistence": "<1.3",
         "phpdocumentor/reflection-docblock": "<5.2|>=7",
         "phpdocumentor/type-resolver": "<1.5.1",
+        "symfony/asset": "<8.2",
         "symfony/asset-mapper": "<8.2",
         "symfony/console": "<8.1.2",
         "symfony/form": "<7.4",
