+++ src/Symfony/Component/Asset/Resources/config/assets.php
