+++ src/Symfony/Component/Asset/composer.json
@@ -19,9 +19,14 @@
         "php": ">=8.4.1"
     },
     "require-dev": {
+        "symfony/config": "^7.4|^8.0",
+        "symfony/dependency-injection": "^8.2",
         "symfony/http-client": "^7.4|^8.0",
         "symfony/http-foundation": "^7.4|^8.0",
         "symfony/http-kernel": "^7.4|^8.0"
+    },
+    "conflict": {
+        "symfony/dependency-injection": "<8.2"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Asset\\": "" },
